requirements.txt
fastapi
uvicorn
python-multipart
streamlit
requests
langchain
langchain-community
langchain-core
sentence-transformers
chromadb
groq
openai-whisper
numpy
scipy
pypdf
pytesseract
pillow
mlflow
prometheus-fastapi-instrumentator
transformers
torch
python-dotenv